Fix BusStops and split ArrivalTimes
